Rename schema.builder.loader package
